projetoBack/requirements.txt
Flask
mysql-connector-python
python-dotenv
wikipedia
requests
langdetect
deep-translator
spacy
scikit-learn
cachetools